// src/types/form.ts
export interface FormDataType {
  name: string;
  email: string;
  phone: string;
  college: string;

  projectType: string;
  platform: string;

  authentication: string[];
  features: string[];
  requirements: string[];

  projectTitle: string;
  deadline: string;
  referenceLink: string;
  description: string;
}